+++ lizard_map/templatetags/map.py
@@ -10,3 +10,10 @@ def background_map(background_map):
     The div tag is picked up by lizard-wms.js
     """
     return {'background_map': background_map}
+
+
+@register.inclusion_tag('lizard_map/tag_animation_control_panel.html')
+def animation_control_panel(workspace):
+    """
+    """
+    return {'workspace': workspace}